Some clang-tidy tips

+++ src/SeisSol/ReceiverDB.h
@@ -19,7 +19,7 @@ class ReceiverDB {
   [[nodiscard]] size_t numberOfReceivers(size_t fsn = 1) const;
 
   private:
-  std::map<int, std::map<int, SeisSol::Receiver>> receivers;
+  std::map<size_t, std::map<size_t, SeisSol::Receiver>> receivers;
   const IO::ReceiverReader receiverReader;
   std::string directory;
   std::string prefix;
